+++ examples/Project/tests/stubs/ReturnValue.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PackageX;
+
+class ReturnValue
+{
+    public function foo() : MyObject
+    {
+    }
+}